feat(maaend): 开发模式手动中止任务时保留游戏进程
