+++ app/utils/__init__.py
@@ -33,7 +33,7 @@ from .security import (
     format_exception_reason,
     sanitize_log_message,
 )
-from .supervision import is_supervised
+from .supervision import is_backend_dev_mode, is_supervised
 
 _LAZY_EXPORTS = {
     "LogMonitor": (".LogMonitor", "LogMonitor"),
@@ -135,6 +135,7 @@ __all__ = [
     "dpapi_decrypt",
     "format_exception_reason",
     "sanitize_log_message",
+    "is_backend_dev_mode",
     "is_supervised",
     "resource_path",
     "strptime",
